refactor(mcp): split 2417-LOC server.go into focused handler files
server.go held the Server type, wiring, and every tool/resource/prompt handler
in one 2417-line file. Move the handlers out, by kind, into sibling files in the
same package — pure code movement, no behavior change:

- server.go (680): Server type, options, NewServer, middleware, Serve*,
  register*, repo-path/cache helpers, ensureCGPService, timeNowUTC
- server_tools.go (1082): all 17 tool handlers + parseRemoteURL
- server_resources.go (340): 5 resource handlers
- server_prompts.go (344): 7 prompt handlers

No exported or unexported identifier changed; same package, so no call sites move.
Existing mcp tests pass unchanged, pinning parity.
