+++ internal/mcp/server.go
@@ -13,7 +13,6 @@ import (
 	mcpmw "go.klarlabs.de/mcp/middleware"
 	"go.klarlabs.de/mcp/transport"
 
-	"github.com/relicta-tech/relicta/v4/internal/cgp"
 	"github.com/relicta-tech/relicta/v4/internal/cgp/evaluator"
 	"github.com/relicta-tech/relicta/v4/internal/cgp/policy"
 	cgpprotocol "github.com/relicta-tech/relicta/v4/internal/cgp/protocol"
@@ -22,7 +21,6 @@ import (
 	"github.com/relicta-tech/relicta/v4/internal/domain/release"
 	relictaerrors "github.com/relicta-tech/relicta/v4/internal/errors"
 	"github.com/relicta-tech/relicta/v4/internal/infrastructure/git"
-	cgpsdk "github.com/relicta-tech/relicta/v4/pkg/cgp"
 )
 
 // ConfigReloader is called after relicta_init creates a config file mid-session.
@@ -662,1649 +660,6 @@ func (s *Server) ensureRepoPath(ctx context.Context) string {
 	}
 	return repoPath
 }
-
-// Tool handlers
-
-func (s *Server) handleInit(ctx context.Context, input InitToolInput) (string, error) {
-	// Get current working directory for config placement
-	repoPath := s.ensureRepoPath(ctx)
-
-	// Check for existing config
-	existingConfig, _ := config.FindConfigFile(repoPath)
-	if existingConfig != "" && !input.Force {
-		result := map[string]any{
-			"status":      "exists",
-			"config_file": existingConfig,
-			"message":     "Configuration file already exists. Use force=true to overwrite.",
-		}
-		jsonBytes, _ := json.Marshal(result)
-		return string(jsonBytes), nil
-	}
-
-	// Determine config file name based on format
-	configFile := ".relicta.yaml"
-	format := input.Format
-	if format == "" {
-		format = "yaml"
-	}
-	if format == "json" {
-		configFile = ".relicta.json"
-	}
-
-	// Start with default configuration
-	cfg := config.DefaultConfig()
-
-	// Try to detect repository settings from git
-	if s.gitService != nil {
-		// Get repository info
-		info, err := s.gitService.GetRepositoryInfo(ctx)
-		if err == nil && info.DefaultBranch != "" {
-			cfg.Workflow.AllowedBranches = []string{info.DefaultBranch}
-		}
-
-		// Try to detect remote URL and extract owner/repo
-		remoteURL, err := s.gitService.GetRemoteURL(ctx, "origin")
-		if err == nil && remoteURL != "" {
-			// Parse GitHub/GitLab URL
-			repoURL := parseRemoteURL(remoteURL)
-			if repoURL != "" {
-				cfg.Changelog.RepositoryURL = repoURL
-			}
-		}
-	}
-
-	// Write config file
-	if err := config.WriteConfig(cfg, configFile); err != nil {
-		return "", fmt.Errorf("failed to write config file: %w", err)
-	}
-
-	result := map[string]any{
-		"status":      "created",
-		"config_file": configFile,
-		"format":      format,
-		"message":     fmt.Sprintf("Created %s with default configuration", configFile),
-		"next_steps": []string{
-			"Review and customize your config file",
-			"Set up required environment variables (OPENAI_API_KEY, GITHUB_TOKEN)",
-			"Run 'relicta plan' to analyze your commits",
-		},
-	}
-
-	// Add detected settings info
-	if cfg.Changelog.RepositoryURL != "" {
-		result["detected_repository"] = cfg.Changelog.RepositoryURL
-	}
-	if len(cfg.Workflow.AllowedBranches) > 0 {
-		result["detected_branch"] = cfg.Workflow.AllowedBranches[0]
-	}
-
-	// Hot-reload: reinitialize config and adapter so subsequent commands work
-	// without restarting the MCP server (fixes #83).
-	if s.configReloader != nil {
-		newCfg, newAdapter, err := s.configReloader(ctx)
-		if err != nil {
-			s.logger.Warn("config reload after init failed, tools may require server restart", "error", err)
-			result["reload_warning"] = "Config created but live reload failed. Restart MCP server if tools return not_configured."
-		} else {
-			s.mu.Lock()
-			s.config = newCfg
-			if newAdapter != nil {
-				s.adapter = newAdapter
-			}
-			s.mu.Unlock()
-			if s.cache != nil {
-				s.cache.InvalidateAll()
-			}
-			s.logger.Info("config reloaded after init, tools are now available")
-			result["reloaded"] = true
-		}
-	}
-
-	jsonBytes, _ := json.Marshal(result)
-	return string(jsonBytes), nil
-}
-
-// parseRemoteURL extracts repository URL from a git remote URL.
-func parseRemoteURL(remoteURL string) string {
-	// Handle SSH format: git@github.com:owner/repo.git
-	if len(remoteURL) > 15 && remoteURL[:15] == "git@github.com:" {
-		path := remoteURL[15:]
-		if len(path) > 4 && path[len(path)-4:] == ".git" {
-			path = path[:len(path)-4]
-		}
-		return "https://github.com/" + path
-	}
-
-	// Handle SSH format: git@gitlab.com:owner/repo.git
-	if len(remoteURL) > 15 && remoteURL[:15] == "git@gitlab.com:" {
-		path := remoteURL[15:]
-		if len(path) > 4 && path[len(path)-4:] == ".git" {
-			path = path[:len(path)-4]
-		}
-		return "https://gitlab.com/" + path
-	}
-
-	// Handle HTTPS format
-	if len(remoteURL) > 8 && (remoteURL[:8] == "https://" || remoteURL[:7] == "http://") {
-		url := remoteURL
-		if len(url) > 4 && url[len(url)-4:] == ".git" {
-			url = url[:len(url)-4]
-		}
-		return url
-	}
-
-	return ""
-}
-
-func (s *Server) handleStatus(ctx context.Context, input StatusInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	s.ensureRepoPath(ctx)
-
-	// Use adapter if available (GetStatus uses releaseServices, not releaseRepo)
-	if s.adapter != nil && s.adapter.HasReleaseServices() {
-		status, err := s.adapter.GetStatus(ctx)
-		if err != nil {
-			return toJSONString(map[string]any{
-				"status":  "no_active_release",
-				"message": "No active release found. Run 'relicta plan' to start a new release.",
-			}), nil
-		}
-
-		result := map[string]any{
-			"release_id":  status.ReleaseID,
-			"state":       status.State,
-			"version":     status.Version,
-			"created":     status.CreatedAt,
-			"updated":     status.UpdatedAt,
-			"can_approve": status.CanApprove,
-			"next_action": status.NextAction,
-		}
-
-		if status.ApprovalMsg != "" {
-			result["approval_message"] = status.ApprovalMsg
-		}
-
-		if status.Stale {
-			result["stale"] = true
-			result["warning"] = status.Warning
-		}
-
-		return toJSONString(result), nil
-	}
-
-	// Fallback to direct repository access
-	if s.releaseRepo == nil {
-		return toJSONString(map[string]any{
-			"status":  "not_configured",
-			"message": "No release repository configured. Run 'relicta plan' first.",
-		}), nil
-	}
-
-	releases, err := s.releaseRepo.FindActive(ctx)
-	if err != nil || len(releases) == 0 {
-		return toJSONString(map[string]any{
-			"status":  "no_active_release",
-			"message": "No active release found. Run 'relicta plan' to start a new release.",
-		}), nil
-	}
-
-	rel := releases[0]
-	result := map[string]any{
-		"state":   rel.State().String(),
-		"version": "",
-		"created": rel.CreatedAt().Format("2006-01-02T15:04:05Z07:00"),
-		"updated": rel.UpdatedAt().Format("2006-01-02T15:04:05Z07:00"),
-	}
-
-	if !rel.VersionNext().IsZero() {
-		result["version"] = rel.VersionNext().String()
-	}
-
-	return toJSONString(result), nil
-}
-
-func (s *Server) handlePlan(ctx context.Context, input PlanToolInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	repoPath := s.ensureRepoPath(ctx)
-
-	// Use adapter if available
-	if s.adapter != nil && s.adapter.HasReleaseAnalyzer() {
-		fromRef := ""
-		if input.From != "" && input.From != "auto" {
-			fromRef = input.From
-		}
-
-		planInput := PlanInput{
-			RepositoryPath: repoPath,
-			FromRef:        fromRef,
-			Analyze:        input.Analyze,
-		}
-
-		// Report progress
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 3.0
-			_ = progress.Report(1, &total)
-		}
-
-		output, err := s.adapter.Plan(ctx, planInput)
-		if err != nil {
-			return "", userError(err)
-		}
-
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 3.0
-			_ = progress.Report(2, &total)
-		}
-
-		result := map[string]any{
-			"release_id":      output.ReleaseID,
-			"current_version": output.CurrentVersion,
-			"next_version":    output.NextVersion,
-			"release_type":    output.ReleaseType,
-			"commit_count":    output.CommitCount,
-			"has_breaking":    output.HasBreaking,
-			"has_features":    output.HasFeatures,
-			"has_fixes":       output.HasFixes,
-		}
-
-		// Include commit details when analyze=true
-		if input.Analyze && len(output.Commits) > 0 {
-			commits := make([]map[string]any, 0, len(output.Commits))
-			for _, c := range output.Commits {
-				commit := map[string]any{
-					"sha":     c.SHA,
-					"type":    c.Type,
-					"message": c.Message,
-					"author":  c.Author,
-				}
-				if c.Scope != "" {
-					commit["scope"] = c.Scope
-				}
-				commits = append(commits, commit)
-			}
-			result["commits"] = commits
-		}
-
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 3.0
-			_ = progress.Report(3, &total)
-		}
-
-		s.invalidateCache()
-		return toJSONString(result), nil
-	}
-
-	return toJSONString(map[string]any{
-		"status":  "not_configured",
-		"message": "Run 'relicta mcp serve' with configured dependencies",
-	}), nil
-}
-
-func (s *Server) handleBump(ctx context.Context, input BumpToolInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	repoPath := s.ensureRepoPath(ctx)
-
-	bumpType := input.Level
-	if bumpType == "" {
-		bumpType = "auto"
-	}
-
-	// Use adapter if available
-	if s.adapter != nil && s.adapter.HasReleaseServices() {
-		bumpInput := BumpInput{
-			RepositoryPath: repoPath,
-			BumpType:       bumpType,
-			Version:        input.Version,
-		}
-
-		output, err := s.adapter.Bump(ctx, bumpInput)
-		if err != nil {
-			return "", userError(err)
-		}
-
-		result := map[string]any{
-			"current_version": output.CurrentVersion,
-			"next_version":    output.NextVersion,
-			"bump_type":       output.BumpType,
-			"auto_detected":   output.AutoDetected,
-		}
-
-		if output.TagName != "" {
-			result["tag_name"] = output.TagName
-			result["tag_created"] = output.TagCreated
-		}
-
-		s.invalidateCache()
-		return toJSONString(result), nil
-	}
-
-	return "", errNotConfigured("relicta_bump")
-}
-
-func (s *Server) handleNotes(ctx context.Context, input NotesToolInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	s.ensureRepoPath(ctx)
-
-	// Use adapter if available (GetStatus and Notes both use releaseServices)
-	if s.adapter != nil && s.adapter.HasReleaseServices() {
-		status, err := s.adapter.GetStatus(ctx)
-		if err != nil {
-			return "", fmt.Errorf("no active release: %w", err)
-		}
-
-		// Report progress
-		totalSteps := 3.0
-		if input.AI {
-			totalSteps = 5.0
-		}
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			_ = progress.Report(1, &totalSteps)
-		}
-
-		notesInput := NotesInput{
-			ReleaseID:        status.ReleaseID,
-			UseAI:            input.AI,
-			IncludeChangelog: true,
-		}
-
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			_ = progress.Report(2, &totalSteps)
-		}
-
-		output, err := s.adapter.Notes(ctx, notesInput)
-		if err != nil {
-			return "", userError(err)
-		}
-
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			_ = progress.Report(totalSteps, &totalSteps)
-		}
-
-		result := map[string]any{
-			"summary":      output.Summary,
-			"ai_generated": output.AIGenerated,
-		}
-
-		if output.Changelog != "" {
-			result["changelog"] = output.Changelog
-		}
-
-		s.invalidateCache()
-		return toJSONString(result), nil
-	}
-
-	return "", errNotConfigured("relicta_notes")
-}
-
-func (s *Server) handleEvaluate(ctx context.Context, input EvaluateToolInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	s.ensureRepoPath(ctx)
-
-	// Use adapter for full governance evaluation if available (GetStatus uses releaseServices)
-	if s.adapter != nil && s.adapter.HasGovernanceService() && s.adapter.HasReleaseServices() {
-		status, err := s.adapter.GetStatus(ctx)
-		if err != nil {
-			return "", fmt.Errorf("no active release: %w", err)
-		}
-
-		// Report progress
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 4.0
-			_ = progress.Report(1, &total)
-		}
-
-		evalInput := EvaluateInput{
-			ReleaseID:      status.ReleaseID,
-			IncludeHistory: true,
-		}
-
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 4.0
-			_ = progress.Report(2, &total)
-		}
-
-		output, err := s.adapter.Evaluate(ctx, evalInput)
-		if err != nil {
-			return "", userError(err)
-		}
-
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 4.0
-			_ = progress.Report(4, &total)
-		}
-
-		return toJSONString(map[string]any{
-			"decision":         output.Decision,
-			"risk_score":       output.RiskScore,
-			"severity":         output.Severity,
-			"can_auto_approve": output.CanAutoApprove,
-			"required_actions": output.RequiredActions,
-			"risk_factors":     output.RiskFactors,
-			"rationale":        output.Rationale,
-		}), nil
-	}
-
-	// Fallback to basic risk calculation
-	if s.riskCalc == nil {
-		return "", fmt.Errorf("risk calculator not configured")
-	}
-
-	proposal := cgp.NewProposal(
-		cgp.Actor{
-			Kind: cgp.ActorKindAgent,
-			ID:   "mcp-client",
-			Name: "MCP Agent",
-		},
-		cgp.ProposalScope{
-			Repository:  "unknown",
-			CommitRange: "HEAD~5..HEAD",
-		},
-		cgp.ProposalIntent{
-			Summary:    "Release evaluation via MCP",
-			Confidence: 0.8,
-		},
-	)
-
-	assessment, err := s.riskCalc.Calculate(ctx, proposal, nil)
-	if err != nil {
-		return "", userError(err)
-	}
-
-	return toJSONString(map[string]any{
-		"score":    assessment.Score,
-		"severity": string(assessment.Severity),
-		"summary":  assessment.Summary,
-		"factors":  assessment.Factors,
-	}), nil
-}
-
-func (s *Server) handleApprove(ctx context.Context, input ApproveToolInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	s.ensureRepoPath(ctx)
-
-	// Use adapter if available (GetStatus and Approve both use releaseServices)
-	if s.adapter != nil && s.adapter.HasReleaseServices() {
-		status, err := s.adapter.GetStatus(ctx)
-		if err != nil {
-			return "", fmt.Errorf("no active release: %w", err)
-		}
-
-		// Enforce actor autonomy budget before privileged operation.
-		if err := s.checkBudget(ctx, policy.Operation{
-			Tool:        "relicta_approve",
-			BlastRadius: blastRadiusForRiskScore(status.RiskScore),
-			RiskScore:   status.RiskScore,
-		}); err != nil {
-			return "", err
-		}
-
-		// Elicitation: for major version releases or releases with breaking changes,
-		// ask the agent for explicit confirmation before proceeding. This leverages
-		// MCP elicitation to ensure the human operator acknowledges the release.
-		if elicitor := mcp.ElicitFromContext(ctx); elicitor != nil {
-			result, elicitErr := elicitor.Elicit(ctx, &mcp.ElicitRequest{
-				Message: fmt.Sprintf(
-					"Approve release %s (version %s) for publishing?",
-					status.ReleaseID, status.Version,
-				),
-			})
-			if elicitErr != nil {
-				s.logger.Debug("elicitation unavailable, proceeding with approval", "error", elicitErr)
-			} else if result.Action != "accept" {
-				return toJSONString(map[string]any{
-					"approved": false,
-					"reason":   "release approval declined",
-				}), nil
-			}
-		}
-
-		approveInput := ApproveInput{
-			ReleaseID:   status.ReleaseID,
-			ApprovedBy:  "mcp-agent",
-			AutoApprove: true,
-			EditedNotes: input.Notes,
-		}
-
-		output, err := s.adapter.Approve(ctx, approveInput)
-		if err != nil {
-			return "", userError(err)
-		}
-
-		s.invalidateCache()
-		return toJSONString(map[string]any{
-			"approved":    output.Approved,
-			"approved_by": output.ApprovedBy,
-			"version":     output.Version,
-		}), nil
-	}
-
-	return "", errNotConfigured("relicta_approve")
-}
-
-func (s *Server) handlePublish(ctx context.Context, input PublishToolInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	s.ensureRepoPath(ctx)
-
-	// Use adapter if available (GetStatus and Publish both use releaseServices)
-	if s.adapter != nil && s.adapter.HasReleaseServices() {
-		status, err := s.adapter.GetStatus(ctx)
-		if err != nil {
-			return "", fmt.Errorf("no active release: %w", err)
-		}
-
-		// Enforce actor autonomy budget before publish.
-		if err := s.checkBudget(ctx, policy.Operation{
-			Tool:        "relicta_publish",
-			BlastRadius: blastRadiusForRiskScore(status.RiskScore),
-			RiskScore:   status.RiskScore,
-		}); err != nil {
-			return "", err
-		}
-
-		// Report progress
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 5.0
-			_ = progress.Report(1, &total)
-		}
-
-		publishInput := PublishInput{
-			ReleaseID: status.ReleaseID,
-			DryRun:    input.DryRun,
-			CreateTag: true,
-			PushTag:   !input.DryRun,
-		}
-
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 5.0
-			_ = progress.Report(2, &total)
-		}
-
-		output, err := s.adapter.Publish(ctx, publishInput)
-		if err != nil {
-			return "", userError(err)
-		}
-
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 5.0
-			_ = progress.Report(4, &total)
-		}
-
-		result := map[string]any{
-			"tag_name":    output.TagName,
-			"release_url": output.ReleaseURL,
-			"dry_run":     input.DryRun,
-		}
-
-		if len(output.PluginResults) > 0 {
-			plugins := make([]map[string]any, 0, len(output.PluginResults))
-			for _, pr := range output.PluginResults {
-				plugins = append(plugins, map[string]any{
-					"plugin":  pr.PluginName,
-					"hook":    pr.Hook,
-					"success": pr.Success,
-					"message": pr.Message,
-				})
-			}
-			result["plugin_results"] = plugins
-		}
-
-		if progress := mcp.ProgressFromContext(ctx); progress != nil {
-			total := 5.0
-			_ = progress.Report(5, &total)
-		}
-
-		s.invalidateCache()
-		return toJSONString(result), nil
-	}
-
-	return "", errNotConfigured("relicta_publish")
-}
-
-func (s *Server) handleCancel(ctx context.Context, input CancelToolInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	s.ensureRepoPath(ctx)
-
-	// Use adapter if available (GetStatus uses releaseServices, Cancel uses releaseRepo)
-	if s.adapter != nil && s.adapter.HasReleaseServices() && s.adapter.HasReleaseRepository() {
-		status, err := s.adapter.GetStatus(ctx)
-		if err != nil {
-			return "", fmt.Errorf("no active release to cancel: %w", err)
-		}
-
-		// Check if release can be canceled
-		if status.State == "published" {
-			return "", fmt.Errorf("cannot cancel a published release")
-		}
-		if status.State == "publishing" && !input.Force {
-			return "", fmt.Errorf("cannot cancel during publishing - wait for completion or use force=true")
-		}
-		if status.State == "failed" || status.State == "canceled" {
-			return toJSONString(map[string]any{
-				"release_id": status.ReleaseID,
-				"state":      status.State,
-				"message":    "release is already in terminal state - use reset to start fresh",
-			}), nil
-		}
-
-		// Enforce actor autonomy budget on cancel.
-		if err := s.checkBudget(ctx, policy.Operation{
-			Tool:        "relicta_cancel",
-			BlastRadius: blastRadiusForRiskScore(status.RiskScore),
-			RiskScore:   status.RiskScore,
-		}); err != nil {
-			return "", err
-		}
-
-		reason := input.Reason
-		if reason == "" {
-			reason = "canceled via MCP"
-		}
-
-		// Cancel the release
-		cancelInput := CancelInput{
-			ReleaseID: status.ReleaseID,
-			Reason:    reason,
-		}
-
-		output, err := s.adapter.Cancel(ctx, cancelInput)
-		if err != nil {
-			return "", userError(err)
-		}
-
-		s.invalidateCache()
-		return toJSONString(map[string]any{
-			"release_id":     output.ReleaseID,
-			"previous_state": output.PreviousState,
-			"new_state":      output.NewState,
-			"reason":         reason,
-			"message":        "release canceled successfully",
-		}), nil
-	}
-
-	return "", errNotConfigured("relicta_cancel")
-}
-
-func (s *Server) handleReset(ctx context.Context, input ResetToolInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	s.ensureRepoPath(ctx)
-
-	// Use adapter if available (GetStatus uses releaseServices, Reset uses releaseRepo)
-	if s.adapter != nil && s.adapter.HasReleaseServices() && s.adapter.HasReleaseRepository() {
-		status, err := s.adapter.GetStatus(ctx)
-		if err != nil {
-			return toJSONString(map[string]any{
-				"message": "no active release found - nothing to reset",
-			}), nil
-		}
-
-		// Check if release can be reset
-		if status.State == "published" {
-			return "", fmt.Errorf("published releases cannot be reset - run 'relicta plan' to start a new release")
-		}
-		if status.State == "publishing" && !input.Force {
-			return "", fmt.Errorf("cannot reset during publishing - wait for completion or use force=true")
-		}
-
-		// For in-progress releases that aren't failed/canceled, suggest cancel first
-		if status.State != "failed" && status.State != "canceled" && !input.Force {
-			return toJSONString(map[string]any{
-				"release_id": status.ReleaseID,
-				"state":      status.State,
-				"message":    "release is in progress - use cancel first, or force=true to delete",
-			}), nil
-		}
-
-		// Enforce actor autonomy budget on reset (destructive).
-		if err := s.checkBudget(ctx, policy.Operation{
-			Tool:        "relicta_reset",
-			BlastRadius: blastRadiusForRiskScore(status.RiskScore),
-			RiskScore:   status.RiskScore,
-		}); err != nil {
-			return "", err
-		}
-
-		// Reset (delete) the release
-		resetInput := ResetInput{
-			ReleaseID: status.ReleaseID,
-		}
-
-		output, err := s.adapter.Reset(ctx, resetInput)
-		if err != nil {
-			return "", userError(err)
-		}
-
-		s.invalidateCache()
-		return toJSONString(map[string]any{
-			"release_id":     output.ReleaseID,
-			"previous_state": output.PreviousState,
-			"deleted":        output.Deleted,
-			"message":        "release reset successfully - run 'relicta plan' to start fresh",
-		}), nil
-	}
-
-	return "", errNotConfigured("relicta_reset")
-}
-
-// --- Specialized AI Agent Tool Handlers ---
-
-func (s *Server) handleBlastRadius(ctx context.Context, input BlastRadiusToolInput) (string, error) {
-	if s.adapter == nil || !s.adapter.HasBlastService() {
-		return toJSONString(map[string]any{
-			"status":  "not_configured",
-			"message": "Blast radius service not configured. This tool requires monorepo analysis to be enabled.",
-		}), nil
-	}
-
-	// Report progress
-	if progress := mcp.ProgressFromContext(ctx); progress != nil {
-		total := 4.0
-		_ = progress.Report(1, &total)
-	}
-
-	blastInput := BlastRadiusInput{
-		FromRef:           input.From,
-		ToRef:             input.To,
-		IncludeTransitive: input.Transitive,
-		GenerateGraph:     input.Graph,
-		PackagePaths:      input.PackagePaths,
-	}
-
-	if progress := mcp.ProgressFromContext(ctx); progress != nil {
-		total := 4.0
-		_ = progress.Report(2, &total)
-	}
-
-	output, err := s.adapter.BlastRadius(ctx, blastInput)
-	if err != nil {
-		return "", userError(err)
-	}
-
-	if progress := mcp.ProgressFromContext(ctx); progress != nil {
-		total := 4.0
-		_ = progress.Report(4, &total)
-	}
-
-	result := map[string]any{
-		"total_packages":             output.TotalPackages,
-		"directly_affected":          output.DirectlyAffected,
-		"transitively_affected":      output.TransitivelyAffected,
-		"packages_requiring_release": output.PackagesRequiringRelease,
-		"risk_level":                 output.RiskLevel,
-		"total_files_changed":        output.TotalFilesChanged,
-		"total_insertions":           output.TotalInsertions,
-		"total_deletions":            output.TotalDeletions,
-	}
-
-	if len(output.RiskFactors) > 0 {
-		result["risk_factors"] = output.RiskFactors
-	}
-
-	if len(output.Impacts) > 0 {
-		impacts := make([]map[string]any, 0, len(output.Impacts))
-		for _, impact := range output.Impacts {
-			impactMap := map[string]any{
-				"package_name":     impact.PackageName,
-				"package_path":     impact.PackagePath,
-				"package_type":     impact.PackageType,
-				"impact_level":     impact.ImpactLevel,
-				"risk_score":       impact.RiskScore,
-				"requires_release": impact.RequiresRelease,
-				"changed_files":    impact.ChangedFiles,
-			}
-			if impact.ReleaseType != "" {
-				impactMap["release_type"] = impact.ReleaseType
-			}
-			if len(impact.SuggestedActions) > 0 {
-				impactMap["suggested_actions"] = impact.SuggestedActions
-			}
-			impacts = append(impacts, impactMap)
-		}
-		result["impacts"] = impacts
-	}
-
-	if output.DependencyGraph != nil {
-		result["dependency_graph"] = map[string]any{
-			"nodes": output.DependencyGraph.Nodes,
-			"edges": output.DependencyGraph.Edges,
-		}
-	}
-
-	return toJSONString(result), nil
-}
-
-func (s *Server) handleInferVersion(ctx context.Context, input InferVersionToolInput) (string, error) {
-	if s.adapter == nil || !s.adapter.HasReleaseAnalyzer() {
-		return toJSONString(map[string]any{
-			"status":  "not_configured",
-			"message": "Release analyzer not configured. Run 'relicta mcp serve' with configured dependencies.",
-		}), nil
-	}
-
-	// Report progress
-	if progress := mcp.ProgressFromContext(ctx); progress != nil {
-		total := 2.0
-		_ = progress.Report(1, &total)
-	}
-
-	inferInput := InferVersionInput{
-		FromRef:     input.From,
-		ToRef:       input.To,
-		IncludeRisk: input.IncludeRisk,
-	}
-
-	output, err := s.adapter.InferVersion(ctx, inferInput)
-	if err != nil {
-		return "", userError(err)
-	}
-
-	if progress := mcp.ProgressFromContext(ctx); progress != nil {
-		total := 2.0
-		_ = progress.Report(2, &total)
-	}
-
-	result := map[string]any{
-		"current_version": output.CurrentVersion,
-		"next_version":    output.NextVersion,
-		"bump_type":       output.BumpType,
-		"has_breaking":    output.HasBreaking,
-		"has_features":    output.HasFeatures,
-		"has_fixes":       output.HasFixes,
-		"commit_count":    output.CommitCount,
-		"confidence":      output.Confidence,
-	}
-
-	if len(output.Rationale) > 0 {
-		result["rationale"] = output.Rationale
-	}
-
-	if input.IncludeRisk {
-		result["risk_score"] = output.RiskScore
-		result["risk_severity"] = output.RiskSeverity
-	}
-
-	return toJSONString(result), nil
-}
-
-func (s *Server) handleSummarizeDiff(ctx context.Context, input SummarizeDiffToolInput) (string, error) {
-	if s.adapter == nil || !s.adapter.HasReleaseAnalyzer() {
-		return toJSONString(map[string]any{
-			"status":  "not_configured",
-			"message": "Release analyzer not configured. Run 'relicta mcp serve' with configured dependencies.",
-		}), nil
-	}
-
-	summarizeInput := SummarizeDiffInput{
-		FromRef:   input.From,
-		ToRef:     input.To,
-		Audience:  input.Audience,
-		MaxLength: input.MaxLength,
-	}
-
-	output, err := s.adapter.SummarizeDiff(ctx, summarizeInput)
-	if err != nil {
-		return "", userError(err)
-	}
-
-	result := map[string]any{
-		"summary":         output.Summary,
-		"audience":        output.Audience,
-		"ai_generated":    output.AIGenerated,
-		"character_count": output.CharacterCount,
-	}
-
-	if len(output.Highlights) > 0 {
-		result["highlights"] = output.Highlights
-	}
-	if len(output.Categories) > 0 {
-		result["categories"] = output.Categories
-	}
-	if output.Signals != nil {
-		result["signals"] = output.Signals
-		// Flag the structured-AI path so agents and CI can branch on it.
-		result["structured"] = true
-	}
-
-	return toJSONString(result), nil
-}
-
-func (s *Server) handleValidateRelease(ctx context.Context, input ValidateReleaseToolInput) (string, error) {
-	// Ensure consistent repository path (fixes issue #35)
-	s.ensureRepoPath(ctx)
-
-	// Get release ID from input or active release
-	releaseID := input.ReleaseID
-	if releaseID == "" && s.adapter != nil && s.adapter.HasReleaseRepository() {
-		status, err := s.adapter.GetStatus(ctx)
-		if err == nil {
-			releaseID = status.ReleaseID
-		}
-	}
-
-	validateInput := ValidateReleaseInput{
-		ReleaseID:       releaseID,
-		CheckGit:        input.CheckGit,
-		CheckPlugins:    input.CheckPlugins,
-		CheckGovernance: input.CheckGovernance,
-		Checks:          input.Checks,
-	}
-
-	// Use adapter if available, otherwise run minimal checks
-	if s.adapter != nil {
-		output, err := s.adapter.ValidateRelease(ctx, validateInput)
-		if err != nil {
-			return "", userError(err)
-		}
-
-		result := map[string]any{
-			"valid":          output.Valid,
-			"can_proceed":    output.CanProceed,
-			"recommendation": output.Recommendation,
-		}
-
-		if len(output.Checks) > 0 {
-			checks := make([]map[string]any, 0, len(output.Checks))
-			for _, check := range output.Checks {
-				checkMap := map[string]any{
-					"name":   check.Name,
-					"status": check.Status,
-				}
-				if check.Message != "" {
-					checkMap["message"] = check.Message
-				}
-				checks = append(checks, checkMap)
-			}
-			result["checks"] = checks
-		}
-
-		if len(output.BlockingIssues) > 0 {
-			result["blocking_issues"] = output.BlockingIssues
-		}
-
-		if len(output.Warnings) > 0 {
-			result["warnings"] = output.Warnings
-		}
-
-		return toJSONString(result), nil
-	}
-
-	// Minimal validation without adapter
-	return toJSONString(map[string]any{
-		"valid":          true,
-		"can_proceed":    true,
-		"recommendation": "Basic validation passed. Full validation requires configured dependencies.",
-		"checks": []map[string]any{
-			{"name": "basic", "status": "passed", "message": "Basic checks passed"},
-		},
-	}), nil
-}
-
-// Resource handlers
-
-func (s *Server) handleResourceState(ctx context.Context, uri string, params map[string]string) (*mcp.ResourceContent, error) {
-	// Check cache first
-	if s.cache != nil {
-		if cached := s.cache.Get(uri); cached != nil {
-			s.logger.Debug("cache hit", "uri", uri)
-			if len(cached.Contents) > 0 {
-				return &mcp.ResourceContent{
-					URI:      cached.Contents[0].URI,
-					MimeType: cached.Contents[0].MIMEType,
-					Text:     cached.Contents[0].Text,
-				}, nil
-			}
-		}
-	}
-
-	if s.releaseRepo == nil {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     `{"status": "no release repository configured"}`,
-		}, nil
-	}
-
-	releases, err := s.releaseRepo.FindActive(ctx)
-	if err != nil || len(releases) == 0 {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     `{"status": "no active release"}`,
-		}, nil
-	}
-
-	rel := releases[0]
-	version := ""
-	if !rel.VersionNext().IsZero() {
-		version = rel.VersionNext().String()
-	}
-
-	content := fmt.Sprintf(`{
-  "state": %q,
-  "version": %q,
-  "created_at": %q,
-  "updated_at": %q
-}`, rel.State().String(), version, rel.CreatedAt().Format("2006-01-02T15:04:05Z07:00"), rel.UpdatedAt().Format("2006-01-02T15:04:05Z07:00"))
-
-	result := &mcp.ResourceContent{
-		URI:      uri,
-		MimeType: "application/json",
-		Text:     content,
-	}
-
-	// Cache the result
-	if s.cache != nil {
-		s.cache.Set(uri, &ReadResourceResult{
-			Contents: []ResourceContent{{URI: uri, MIMEType: "application/json", Text: content}},
-		})
-	}
-
-	return result, nil
-}
-
-func (s *Server) handleResourceConfig(ctx context.Context, uri string, params map[string]string) (*mcp.ResourceContent, error) {
-	if s.config == nil {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     `{"status": "no configuration loaded"}`,
-		}, nil
-	}
-
-	productName := s.config.Changelog.ProductName
-	if productName == "" {
-		productName = "Relicta"
-	}
-
-	content := fmt.Sprintf(`{
-  "product_name": %q,
-  "ai_enabled": %t,
-  "ai_provider": %q,
-  "versioning_strategy": %q
-}`, productName, s.config.AI.Enabled, s.config.AI.Provider, s.config.Versioning.Strategy)
-
-	return &mcp.ResourceContent{
-		URI:      uri,
-		MimeType: "application/json",
-		Text:     content,
-	}, nil
-}
-
-func (s *Server) handleResourceCommits(ctx context.Context, uri string, params map[string]string) (*mcp.ResourceContent, error) {
-	if s.releaseRepo == nil {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     `{"status": "no release repository configured"}`,
-		}, nil
-	}
-
-	releases, err := s.releaseRepo.FindActive(ctx)
-	if err != nil || len(releases) == 0 {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     `{"status": "no active release", "commits": []}`,
-		}, nil
-	}
-
-	rel := releases[0]
-	plan := release.GetPlan(rel)
-	if plan == nil {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     `{"status": "no plan available", "commits": []}`,
-		}, nil
-	}
-
-	// Check if changeset is loaded
-	if !plan.HasChangeSet() {
-		content := fmt.Sprintf(`{
-  "status": "changeset not loaded",
-  "changeset_id": %q,
-  "release_type": %q,
-  "current_version": %q,
-  "next_version": %q,
-  "commits": []
-}`, plan.ChangeSetID, plan.ReleaseType, plan.CurrentVersion.String(), plan.NextVersion.String())
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     content,
-		}, nil
-	}
-
-	// Get commits from changeset
-	changeSet := plan.GetChangeSet()
-	commits := changeSet.Commits()
-
-	// Build commits array
-	commitList := make([]map[string]any, 0, len(commits))
-	for _, c := range commits {
-		commit := map[string]any{
-			"sha":      c.ShortHash(),
-			"full_sha": c.Hash(),
-			"type":     string(c.Type()),
-			"subject":  c.Subject(),
-			"author":   c.Author(),
-			"date":     c.Date().Format("2006-01-02T15:04:05Z07:00"),
-			"breaking": c.IsBreaking(),
-		}
-		if c.Scope() != "" {
-			commit["scope"] = c.Scope()
-		}
-		commitList = append(commitList, commit)
-	}
-
-	result := map[string]any{
-		"status":          "ok",
-		"changeset_id":    string(plan.ChangeSetID),
-		"release_type":    string(plan.ReleaseType),
-		"current_version": plan.CurrentVersion.String(),
-		"next_version":    plan.NextVersion.String(),
-		"commit_count":    len(commits),
-		"commits":         commitList,
-	}
-
-	jsonBytes, err := json.MarshalIndent(result, "", "  ")
-	if err != nil {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     fmt.Sprintf(`{"status": "error", "error": %q}`, err.Error()),
-		}, nil
-	}
-
-	return &mcp.ResourceContent{
-		URI:      uri,
-		MimeType: "application/json",
-		Text:     string(jsonBytes),
-	}, nil
-}
-
-func (s *Server) handleResourceChangelog(ctx context.Context, uri string, params map[string]string) (*mcp.ResourceContent, error) {
-	if s.releaseRepo == nil {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "text/markdown",
-			Text:     "# Changelog\n\nNo release repository configured.",
-		}, nil
-	}
-
-	releases, err := s.releaseRepo.FindActive(ctx)
-	if err != nil || len(releases) == 0 {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "text/markdown",
-			Text:     "# Changelog\n\nNo active release found. Run `relicta plan` to start a new release.",
-		}, nil
-	}
-
-	rel := releases[0]
-	notes := rel.Notes()
-
-	if notes == nil {
-		version := ""
-		if !rel.VersionNext().IsZero() {
-			version = rel.VersionNext().String()
-		}
-
-		content := fmt.Sprintf("# Changelog\n\nNo changelog generated yet for version %s.\n\nRun `relicta notes` to generate release notes.", version)
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "text/markdown",
-			Text:     content,
-		}, nil
-	}
-
-	changelog := notes.Text
-	if changelog == "" {
-		changelog = "# Release Notes\n\nNo content available."
-	}
-
-	return &mcp.ResourceContent{
-		URI:      uri,
-		MimeType: "text/markdown",
-		Text:     changelog,
-	}, nil
-}
-
-func (s *Server) handleResourceRiskReport(ctx context.Context, uri string, params map[string]string) (*mcp.ResourceContent, error) {
-	if s.releaseRepo == nil {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     `{"status": "no release repository configured"}`,
-		}, nil
-	}
-
-	releases, err := s.releaseRepo.FindActive(ctx)
-	if err != nil || len(releases) == 0 {
-		return &mcp.ResourceContent{
-			URI:      uri,
-			MimeType: "application/json",
-			Text:     `{"status": "no active release"}`,
-		}, nil
-	}
-
-	rel := releases[0]
-
-	// Try to get risk assessment from adapter if available
-	if s.adapter != nil && s.adapter.HasGovernanceService() {
-		evalInput := EvaluateInput{
-			ReleaseID:      string(rel.ID()),
-			IncludeHistory: false,
-		}
-
-		output, err := s.adapter.Evaluate(ctx, evalInput)
-		if err == nil {
-			result := map[string]any{
-				"status":           "ok",
-				"decision":         output.Decision,
-				"risk_score":       output.RiskScore,
-				"severity":         output.Severity,
-				"can_auto_approve": output.CanAutoApprove,
-				"required_actions": output.RequiredActions,
-				"risk_factors":     output.RiskFactors,
-				"rationale":        output.Rationale,
-			}
-
-			jsonBytes, err := json.MarshalIndent(result, "", "  ")
-			if err == nil {
-				return &mcp.ResourceContent{
-					URI:      uri,
-					MimeType: "application/json",
-					Text:     string(jsonBytes),
-				}, nil
-			}
-		}
-	}
-
-	// Fallback: Use basic risk calculation if available
-	if s.riskCalc != nil {
-		proposal := cgp.NewProposal(
-			cgp.Actor{
-				Kind: cgp.ActorKindAgent,
-				ID:   "mcp-resource-reader",
-				Name: "MCP Resource Reader",
-			},
-			cgp.ProposalScope{
-				Repository:  rel.RepoID(),
-				CommitRange: "HEAD~5..HEAD",
-			},
-			cgp.ProposalIntent{
-				Summary:    "Risk assessment for active release",
-				Confidence: 0.8,
-			},
-		)
-
-		assessment, err := s.riskCalc.Calculate(ctx, proposal, nil)
-		if err == nil {
-			result := map[string]any{
-				"status":   "ok",
-				"score":    assessment.Score,
-				"severity": string(assessment.Severity),
-				"summary":  assessment.Summary,
-				"factors":  assessment.Factors,
-			}
-
-			jsonBytes, err := json.MarshalIndent(result, "", "  ")
-			if err == nil {
-				return &mcp.ResourceContent{
-					URI:      uri,
-					MimeType: "application/json",
-					Text:     string(jsonBytes),
-				}, nil
-			}
-		}
-	}
-
-	return &mcp.ResourceContent{
-		URI:      uri,
-		MimeType: "application/json",
-		Text:     `{"status": "no risk assessment available", "hint": "Run 'relicta evaluate' to perform risk assessment"}`,
-	}, nil
-}
-
-// Prompt handlers
-
-func (s *Server) handlePromptReleaseSummary(ctx context.Context, args map[string]string) (*mcp.PromptResult, error) {
-	style := "brief"
-	if v, ok := args["style"]; ok && v != "" {
-		style = v
-	}
-
-	var content string
-	switch style {
-	case "detailed":
-		content = `You are a release manager. Provide a detailed summary of the upcoming release including:
-- All changes categorized by type (features, fixes, breaking changes)
-- Impact analysis
-- Risk assessment
-- Recommended actions before release`
-	case "technical":
-		content = `You are a technical writer. Provide a technical summary of the release including:
-- API changes and their signatures
-- Configuration changes
-- Migration requirements
-- Performance implications`
-	default:
-		content = `You are a release manager. Provide a brief summary of the upcoming release including:
-- Key highlights (1-3 bullet points)
-- Version number
-- Release readiness status`
-	}
-
-	return &mcp.PromptResult{
-		Description: "Release summary prompt",
-		Messages: []mcp.PromptMessage{
-			{Role: "user", Content: mcp.TextContent{Type: "text", Text: content}},
-		},
-	}, nil
-}
-
-func (s *Server) handlePromptRiskAnalysis(ctx context.Context, args map[string]string) (*mcp.PromptResult, error) {
-	content := `You are a release risk analyst using the Change Governance Protocol (CGP).
-
-Analyze the current release and provide:
-1. Overall risk score (0.0 - 1.0) with justification
-2. Individual risk factors:
-   - API changes impact
-   - Dependency changes
-   - Blast radius (scope of changes)
-   - Security implications
-   - Historical patterns
-3. Recommendations:
-   - Approval recommendation (approve, review, block)
-   - Required actions before release
-   - Suggested reviewers
-
-Base your analysis on the commit history, change analysis, and CGP policies.`
-
-	return &mcp.PromptResult{
-		Description: "Risk analysis prompt",
-		Messages: []mcp.PromptMessage{
-			{Role: "user", Content: mcp.TextContent{Type: "text", Text: content}},
-		},
-	}, nil
-}
-
-func (s *Server) handlePromptCommitReview(ctx context.Context, args map[string]string) (*mcp.PromptResult, error) {
-	focus := "compliance"
-	if v, ok := args["focus"]; ok && v != "" {
-		focus = v
-	}
-
-	var content string
-	switch focus {
-	case "quality":
-		content = `You are a code review expert analyzing commit messages and changes.
-
-Review the commits in this release for quality:
-1. Commit message clarity and completeness
-2. Logical grouping of changes (atomic commits)
-3. Code change quality indicators
-4. Documentation updates for significant changes
-5. Test coverage implications
-
-For each issue found, suggest specific improvements.`
-	case "security":
-		content = `You are a security analyst reviewing commits for potential security implications.
-
-Analyze commits for security concerns:
-1. Sensitive data handling changes
-2. Authentication/authorization modifications
-3. Input validation changes
-4. Dependency updates with known vulnerabilities
-5. Configuration changes affecting security posture
-6. Cryptographic code changes
-
-Flag any commits that require security review before release.`
-	default: // compliance
-		content = `You are a release compliance officer reviewing commits for conventional commit standards.
-
-Analyze each commit for compliance with Conventional Commits specification:
-1. Format: <type>(<scope>): <subject>
-2. Valid types: feat, fix, docs, style, refactor, perf, test, build, ci, chore
-3. Breaking changes properly marked with ! or BREAKING CHANGE footer
-4. Scope relevance and consistency
-5. Subject line length and imperative mood
-
-List non-compliant commits with specific corrections needed.`
-	}
-
-	return &mcp.PromptResult{
-		Description: "Commit review prompt",
-		Messages: []mcp.PromptMessage{
-			{Role: "user", Content: mcp.TextContent{Type: "text", Text: content}},
-		},
-	}, nil
-}
-
-func (s *Server) handlePromptBreakingChanges(ctx context.Context, args map[string]string) (*mcp.PromptResult, error) {
-	content := `You are a technical writer documenting breaking changes for users.
-
-For each breaking change in this release, provide:
-
-1. **Change Summary**: One-line description of what changed
-2. **Reason**: Why this breaking change was necessary
-3. **Impact**: Who is affected and how
-4. **Migration Path**: Step-by-step instructions to adapt
-5. **Code Examples**: Before/after code snippets where applicable
-
-Format the output as a structured breaking changes document suitable for inclusion in release notes.
-
-If there are no breaking changes, confirm this and explain what safeguards prevented them.`
-
-	return &mcp.PromptResult{
-		Description: "Breaking changes documentation prompt",
-		Messages: []mcp.PromptMessage{
-			{Role: "user", Content: mcp.TextContent{Type: "text", Text: content}},
-		},
-	}, nil
-}
-
-func (s *Server) handlePromptMigrationGuide(ctx context.Context, args map[string]string) (*mcp.PromptResult, error) {
-	audience := "developer"
-	if v, ok := args["audience"]; ok && v != "" {
-		audience = v
-	}
-
-	var content string
-	switch audience {
-	case "operator":
-		content = `You are a DevOps engineer writing migration instructions for system operators.
-
-Create a migration guide covering:
-1. **Pre-migration Checklist**
-   - Backup procedures
-   - Downtime requirements
-   - Rollback plan
-
-2. **Infrastructure Changes**
-   - Configuration file updates
-   - Environment variable changes
-   - Database migrations
-
-3. **Deployment Steps**
-   - Ordered deployment sequence
-   - Health check verification
-   - Monitoring updates
-
-4. **Post-migration Verification**
-   - Smoke tests to run
-   - Metrics to monitor
-   - Common issues and solutions
-
-Use clear, actionable language suitable for runbooks.`
-	case "end-user":
-		content = `You are a product manager writing migration notes for end users.
-
-Create user-friendly upgrade instructions:
-1. **What's New**: Key benefits of upgrading
-2. **What's Changed**: User-facing changes to expect
-3. **Action Required**: Steps users need to take
-4. **Getting Help**: Support resources and FAQ
-
-Keep technical jargon minimal. Focus on the user experience impact.`
-	default: // developer
-		content = `You are a senior developer writing migration instructions for other developers.
-
-Create a developer migration guide covering:
-1. **Dependency Updates**
-   - Version requirements
-   - Package manager commands
-
-2. **API Changes**
-   - Changed endpoints/methods
-   - Parameter modifications
-   - Response format changes
-   - Deprecation timeline
-
-3. **Code Migration**
-   - Find/replace patterns
-   - Refactoring steps
-   - Type definition updates
-
-4. **Testing Updates**
-   - Test changes needed
-   - New test patterns
-
-Include code examples for all significant changes.`
-	}
-
-	return &mcp.PromptResult{
-		Description: "Migration guide prompt",
-		Messages: []mcp.PromptMessage{
-			{Role: "user", Content: mcp.TextContent{Type: "text", Text: content}},
-		},
-	}, nil
-}
-
-func (s *Server) handlePromptReleaseAnnouncement(ctx context.Context, args map[string]string) (*mcp.PromptResult, error) {
-	channel := "github"
-	if v, ok := args["channel"]; ok && v != "" {
-		channel = v
-	}
-
-	var content string
-	switch channel {
-	case "blog":
-		content = `You are a technical writer crafting a blog post for a software release.
-
-Write a blog post announcement including:
-1. **Headline**: Compelling title highlighting the main theme
-2. **Introduction**: Context and significance of this release
-3. **Key Features**: Detailed coverage of major additions
-4. **Improvements**: Notable fixes and enhancements
-5. **Technical Deep-dive**: One feature explained in depth
-6. **Getting Started**: How to upgrade or try it
-7. **What's Next**: Roadmap preview
-8. **Acknowledgments**: Contributor recognition
-
-Tone: Professional but engaging, 800-1200 words.`
-	case "social":
-		content = `You are a developer advocate crafting social media announcements.
-
-Create announcements for different platforms:
-
-**Twitter/X (280 chars)**:
-- Hook + key feature + link
-
-**LinkedIn (longer form)**:
-- Professional summary
-- 3 key highlights
-- Call to action
-
-**Mastodon/Dev Community**:
-- Technical focus
-- Code snippet if relevant
-- Community engagement
-
-Use appropriate hashtags and emoji sparingly.`
-	case "email":
-		content = `You are writing a release announcement email for subscribers.
-
-Structure the email:
-1. **Subject Line**: Clear, action-oriented
-2. **Preview Text**: Compelling summary
-3. **Header**: Version number and release date
-4. **Executive Summary**: 2-3 sentence overview
-5. **Highlights**: Bullet points of key changes
-6. **Upgrade Instructions**: Brief steps
-7. **Full Changelog Link**: For detailed information
-8. **Feedback Request**: How to provide input
-
-Keep it scannable with clear visual hierarchy.`
-	default: // github
-		content = `You are writing release notes for a GitHub release.
-
-Structure the release notes:
-1. **Title**: Version number with optional codename
-2. **Summary**: 2-3 sentence release overview
-3. **Highlights**: Top 3-5 changes as bullet points
-4. **What's Changed**: Categorized changes
-   - ✨ Features
-   - 🐛 Bug Fixes
-   - 📚 Documentation
-   - ⚠️ Breaking Changes
-5. **Upgrade Notes**: Critical information for upgrading
-6. **Contributors**: @mention contributors
-
-Use GitHub-flavored markdown with appropriate emoji.`
-	}
-
-	return &mcp.PromptResult{
-		Description: "Release announcement prompt",
-		Messages: []mcp.PromptMessage{
-			{Role: "user", Content: mcp.TextContent{Type: "text", Text: content}},
-		},
-	}, nil
-}
-
-func (s *Server) handlePromptApprovalDecision(ctx context.Context, args map[string]string) (*mcp.PromptResult, error) {
-	content := `You are a release governance advisor helping make approval decisions.
-
-Based on the Change Governance Protocol (CGP) analysis, provide:
-
-1. **Decision Recommendation**: APPROVE, REQUEST_CHANGES, or BLOCK
-   - Clear rationale for your recommendation
-   - Confidence level (high/medium/low)
-
-2. **Risk Assessment Summary**
-   - Overall risk level and score
-   - Top 3 risk factors to consider
-   - Mitigating factors present
-
-3. **Approval Conditions** (if recommending approval)
-   - Required reviewers and their focus areas
-   - Pre-release checks to complete
-   - Monitoring requirements post-release
-
-4. **Blocking Issues** (if recommending block)
-   - Specific issues that must be resolved
-   - Suggested remediation steps
-   - Criteria for re-evaluation
-
-5. **Audit Trail Entry**
-   - Structured summary for governance records
-   - Key decision factors
-   - Timestamp and context
-
-Provide actionable guidance that enables confident decision-making.`
-
-	return &mcp.PromptResult{
-		Description: "Approval decision prompt",
-		Messages: []mcp.PromptMessage{
-			{Role: "user", Content: mcp.TextContent{Type: "text", Text: content}},
-		},
-	}, nil
-}
-
-// =============================================================================
-// CGP Protocol Wire Format Handlers
-// =============================================================================
-
 func (s *Server) ensureCGPService() error {
 	if s.cgpService != nil {
 		return nil
@@ -2317,98 +672,6 @@ func (s *Server) ensureCGPService() error {
 	}
 
 	return fmt.Errorf("CGP protocol service not configured")
-}
-
-func (s *Server) handleCGPPropose(ctx context.Context, input CGPProposeToolInput) (string, error) {
-	if err := s.ensureCGPService(); err != nil {
-		return "", err
-	}
-
-	proposal := &cgpsdk.ChangeProposal{
-		CGPVersion: cgpsdk.ProtocolVersion,
-		Type:       cgpsdk.TypeChangeProposal,
-		ID:         cgp.GenerateProposalID(),
-		Timestamp:  timeNowUTC(),
-		Actor: cgpsdk.Actor{
-			Kind: input.ActorKind,
-			ID:   input.ActorID,
-			Name: input.ActorName,
-		},
-		Scope: cgpsdk.Scope{
-			Repository:  input.Repository,
-			CommitRange: input.CommitRange,
-		},
-		Intent: cgpsdk.Intent{
-			Summary:    input.Summary,
-			Confidence: input.Confidence,
-			Categories: input.Categories,
-		},
-	}
-
-	decision, err := s.cgpService.EvaluateProposal(ctx, proposal)
-	if err != nil {
-		return "", userError(err)
-	}
-
-	// Marshal the decision using the CGP wire format codec.
-	data, err := cgpsdk.Marshal(decision)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal decision: %w", err)
-	}
-
-	return string(data), nil
-}
-
-func (s *Server) handleCGPAuthorize(ctx context.Context, input CGPAuthorizeToolInput) (string, error) {
-	if err := s.ensureCGPService(); err != nil {
-		return "", err
-	}
-
-	now := timeNowUTC()
-	auth := &cgpsdk.ExecutionAuthorization{
-		CGPVersion: cgpsdk.ProtocolVersion,
-		Type:       cgpsdk.TypeExecutionAuthorization,
-		ID:         cgp.GenerateAuthorizationID(),
-		ProposalID: input.ProposalID,
-		DecisionID: input.DecisionID,
-		Timestamp:  now,
-		ApprovedBy: cgpsdk.Actor{
-			Kind: "human",
-			ID:   input.ApproverID,
-		},
-		Version:    input.Version,
-		ValidUntil: now.Add(24 * 60 * 60 * 1e9), // 24 hours
-	}
-
-	if err := s.cgpService.RecordAuthorization(ctx, auth); err != nil {
-		return "", userError(err)
-	}
-
-	data, err := cgpsdk.Marshal(auth)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal authorization: %w", err)
-	}
-
-	return string(data), nil
-}
-
-func (s *Server) handleCGPStatus(ctx context.Context, input CGPStatusToolInput) (string, error) {
-	if err := s.ensureCGPService(); err != nil {
-		return "", err
-	}
-
-	status, err := s.cgpService.GetStatus(ctx, input.ProposalID)
-	if err != nil {
-		return "", userError(err)
-	}
-
-	return toJSONString(map[string]any{
-		"proposalId":    status.ProposalID,
-		"state":         status.State,
-		"proposal":      status.Proposal,
-		"decision":      status.Decision,
-		"authorization": status.Authorization,
-	}), nil
 }
 
 // timeNowUTC returns the current time in UTC. Extracted for testability.
